Enhance root normalization logic and make idempotent.
Previously multiple calls to this would keep adding entries to the `root` array due to calls from `webpack-partial`. Now that we have full control of the config structure we just do the merge logic ourselves.

Also now handle the case where `root` in the webpack configuration is something other than a string.